fix: address type errors
