+++ src/app/api/upload/route.ts
@@ -13,7 +13,13 @@ export const POST = withAuthorization(
   "create",
   async (
     req: Request,
-    { session }: { session?: { user?: { id?: string; role?: string } } },
+    {
+      session,
+      params,
+    }: {
+      params: Promise<Record<string, string>>;
+      session?: { user?: { id?: string; role?: string } };
+    },
   ) => {
     const form = await req.formData();
     const file = form.get("photo") as File | null;
