+++ src/app/api/users/invite/route.ts
@@ -5,7 +5,13 @@ import { NextResponse } from "next/server";
 export const POST = withAuthorization(
   "admin",
   "create",
-  async (req: Request) => {
+  async (
+    req: Request,
+    _ctx: {
+      params: Promise<Record<string, string>>;
+      session?: { user?: { role?: string } };
+    },
+  ) => {
     const { email } = (await req.json()) as { email: string };
     const user = inviteUser(email);
     return NextResponse.json(user);
